Refs #3397. Moved check_buffer function to a common file
