+++ src/c/common_internals.h
@@ -12,8 +12,8 @@
 // See the License for the specific language governing permissions and
 // limitations under the License.
 
-#ifndef _SRC_TYPES_BASIC_INTERNALS_H_
-#define _SRC_TYPES_BASIC_INTERNALS_H_
+#ifndef _SRC_COMMON_INTERNALS_H_
+#define _SRC_COMMON_INTERNALS_H_
 
 #ifdef __cplusplus
 extern "C" {
@@ -22,20 +22,12 @@ extern "C" {
 #include <microcdr/common.h>
 
 // -------------------------------------------------------------------
-//                  INTERNAL SERIALIZATION FUNCTIONS
+//                     INTERNAL UTIL FUNCTIONS
 // -------------------------------------------------------------------
-bool serialize_byte_1(MicroBuffer* buffer, const uint8_t* byte);
-bool serialize_byte_2(MicroBuffer* buffer, Endianness endianness, const uint16_t* bytes);
-bool serialize_byte_4(MicroBuffer* buffer, Endianness endianness, const uint32_t* bytes);
-bool serialize_byte_8(MicroBuffer* buffer, Endianness endianness, const uint64_t* bytes);
-
-bool deserialize_byte_1(MicroBuffer* buffer, uint8_t* byte);
-bool deserialize_byte_2(MicroBuffer* buffer, Endianness endianness, uint16_t* bytes);
-bool deserialize_byte_4(MicroBuffer* buffer, Endianness endianness, uint32_t* bytes);
-bool deserialize_byte_8(MicroBuffer* buffer, Endianness endianness, uint64_t* bytes);
+bool check_buffer(MicroBuffer* mb, const uint32_t bytes);
 
 #ifdef __cplusplus
 }
 #endif
 
-#endif //_SRC_TYPES_BASIC_INTERNALS_H_
+#endif //_SRC_COMMON_INTERNALS_H_
